requirements.txt
# Core Dependencies
streamlit>=1.28.0
numpy>=1.24.0
pandas>=2.0.0
pillow>=10.0.0

# Computer Vision & Image Processing
opencv-python-headless>=4.8.0
opencv-contrib-python-headless>=4.8.0

# Emotion Detection
fer>=22.5.1
tensorflow>=2.13.0

# NLP & Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0
transformers>=4.30.0

# Graph & Network Analysis
networkx>=3.1

# Visualization
plotly>=5.14.0
matplotlib>=3.7.0

# Utilities
python-dateutil>=2.8.2
